EP-2500

+++ eperusteet-amosaa-service/src/main/java/fi/vm/sade/eperusteet/amosaa/dto/teksti/OpintokokonaisuusDto.java
@@ -1,6 +1,7 @@
 package fi.vm.sade.eperusteet.amosaa.dto.teksti;
 
 import fi.vm.sade.eperusteet.amosaa.domain.tutkinnonosa.OpintokokonaisuusTyyppi;
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 import lombok.Data;
@@ -10,7 +11,7 @@ public class OpintokokonaisuusDto {
     private Long id;
     private String nimiKoodi;
     private LokalisoituTekstiDto kuvaus;
-    private Integer laajuus;
+    private BigDecimal laajuus;
     private Integer minimilaajuus;
     private OpintokokonaisuusTyyppi tyyppi;
     private LokalisoituTekstiDto opetuksenTavoiteOtsikko;
